# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APSO Backend API",
    "description": "AI-powered SDLC intelligence backend. MVP scope is read-only extraction and analysis.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/quality/scanners/dry-run": {
      "post": {
        "tags": [
          "quality"
        ],
        "summary": "Scanner Dry Run",
        "operationId": "scanner_dry_run_api_v1_quality_scanners_dry_run_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScannerRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScannerRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quality/release-readiness": {
      "post": {
        "tags": [
          "quality"
        ],
        "summary": "Assess Release Readiness",
        "operationId": "assess_release_readiness_api_v1_quality_release_readiness_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseReadinessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseReadinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReleaseReadinessRequest": {
        "properties": {
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "release_name": {
            "type": "string",
            "title": "Release Name"
          },
          "open_high_risk_findings": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Open High Risk Findings",
            "default": 0
          },
          "failed_pipeline_runs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Failed Pipeline Runs",
            "default": 0
          },
          "stories_without_tests": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stories Without Tests",
            "default": 0
          },
          "unresolved_requirement_gaps": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unresolved Requirement Gaps",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "release_name"
        ],
        "title": "ReleaseReadinessRequest"
      },
      "ReleaseReadinessResponse": {
        "properties": {
          "release_name": {
            "type": "string",
            "title": "Release Name"
          },
          "readiness_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Readiness Score"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blockers"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "release_name",
          "readiness_score",
          "status",
          "blockers"
        ],
        "title": "ReleaseReadinessResponse"
      },
...
      "ScannerFinding": {
        "properties": {
          "scanner": {
            "type": "string",
            "enum": [
              "semgrep",
              "tree_sitter",
              "gitleaks",
              "trivy",
              "owasp_dependency_check"
            ],
            "title": "Scanner"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "evidence": {
            "type": "string",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "scanner",
          "severity",
          "title",
          "evidence"
        ],
        "title": "ScannerFinding"
      },
      "ScannerRunRequest": {
        "properties": {
          "repository_slug": {
            "type": "string",
            "title": "Repository Slug"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "default": "develop"
          },
          "scanner": {
            "type": "string",
            "enum": [
              "semgrep",
              "tree_sitter",
              "gitleaks",
              "trivy",
              "owasp_dependency_check"
            ],
            "title": "Scanner"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "repository_slug",
          "scanner"
        ],
        "title": "ScannerRunRequest"
      },
      "ScannerRunResponse": {
        "properties": {
          "repository_slug": {
            "type": "string",
            "title": "Repository Slug"
          },
          "branch": {
            "type": "string",
            "title": "Branch"
          },
          "scanner": {
            "type": "string",
            "enum": [
              "semgrep",
              "tree_sitter",
              "gitleaks",
              "trivy",
              "owasp_dependency_check"
            ],
            "title": "Scanner"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run"
          },
          "tool_available": {
            "type": "boolean",
            "title": "Tool Available",
            "default": false
          },
          "command": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Command"
          },
          "return_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Code"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ScannerFinding"
            },
            "type": "array",
            "title": "Findings"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "repository_slug",
          "branch",
          "scanner",
          "dry_run",
          "note"
        ],
        "title": "ScannerRunResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}